feat: update buttons on main page
