+++ frontend/components/MainPage/header/index.tsx
@@ -1,19 +1,18 @@
-import { QuestionCircleOutlined, SettingOutlined } from '@ant-design/icons';
-import { Button, Flex } from 'antd';
+import { Flex } from 'antd';
 import { useCallback, useEffect, useState } from 'react';
 
 import { MiddlewareDeploymentStatus } from '@/client';
 import { CardSection } from '@/components/styled/CardSection';
-import { Pages } from '@/enums/Pages';
 import { useBalanceContext } from '@/hooks/useBalanceContext';
 import { useElectronApi } from '@/hooks/useElectronApi';
-import { usePageState } from '@/hooks/usePageState';
 import { useService } from '@/hooks/useService';
 import { useServices } from '@/hooks/useServices';
 
 import { FirstRunModal } from '../modals/FirstRunModal';
 import { AgentButton } from './AgentButton/AgentButton';
 import { AgentHead } from './AgentHead';
+import { AgentProfileButton } from './AgentProfileButton';
+import { AgentSettingsButton } from './AgentSettingsButton';
 
 const useSetupTrayIcon = () => {
   const { isLowBalance } = useBalanceContext();
@@ -37,8 +36,6 @@ const useSetupTrayIcon = () => {
 };
 
 export const MainHeader = () => {
-  const { goto } = usePageState();
-
   const [isFirstRunModalOpen, setIsFirstRunModalOpen] = useState(false);
   const handleModalClose = useCallback(() => setIsFirstRunModalOpen(false), []);
 
@@ -47,25 +44,15 @@ export const MainHeader = () => {
 
   return (
     <CardSection gap={6} padding="8px 24px" justify="space-between">
-      <Flex justify="start" align="center" gap={4}>
+      <Flex justify="start" align="center" gap={8}>
         <AgentHead />
         <AgentButton />
         <FirstRunModal open={isFirstRunModalOpen} onClose={handleModalClose} />
       </Flex>
 
       <Flex gap={8} align="center">
-        <Button
-          type="default"
-          size="large"
-          icon={<QuestionCircleOutlined />}
-          onClick={() => goto(Pages.HelpAndSupport)}
-        />
-        <Button
-          type="default"
-          size="large"
-          icon={<SettingOutlined />}
-          onClick={() => goto(Pages.Settings)}
-        />
+        <AgentSettingsButton />
+        <AgentProfileButton />
       </Flex>
     </CardSection>
   );
