+++ frontend/components/custom-icons/AgentProfile.tsx
@@ -1,0 +1,17 @@
+export const AgentProfileSvg = () => (
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    width="20"
+    height="20"
+    viewBox="0 0 20 20"
+    fill="none"
+  >
+    <path
+      d="M10 6.31965V2.63965H6.32005M0.800049 11.8396H2.64005M17.36 11.8396H19.2M12.76 10.9196V12.7596M7.24005 10.9196V12.7596M4.48005 6.31965H15.52C16.5363 6.31965 17.36 7.14344 17.36 8.15965V15.5196C17.36 16.5359 16.5363 17.3596 15.52 17.3596H4.48005C3.46384 17.3596 2.64005 16.5359 2.64005 15.5196V8.15965C2.64005 7.14344 3.46384 6.31965 4.48005 6.31965Z"
+      stroke="currentColor"
+      stroke-width="1.4"
+      stroke-linecap="round"
+      stroke-linejoin="round"
+    />
+  </svg>
+);
